pred2bq: Replace abseil tempfile creation.
Replaces create_tempfile and create_tempdir calls from abseil's
absltest.TestCase and parameterized.TestCase with equivalent
methods from the tempfile package.

The reason is that the abseil methods require parsing of the
FLAGs variable, which may not be executed if absltest.main() is
not invoked. This can happen when test filtering is performed, e.g.
```
python -m unittest path.to.test
```